+++ examples/command-builder.php
@@ -10,5 +10,5 @@ use function ls\ls;
 #[AsTask(name: 'ls', description: 'Run a sub-process and display information about it')]
 function ls_task(): void
 {
-    run(ls(__DIR__ . '/command-builder')->all()->withWorkingDirectory(__DIR__ . '/command-builder'));
+    run(ls(__DIR__ . '/command-builder')->all());
 }